chore: move the prefix wrapping iterator to its own file
